react router, jwt, reorganized static files
