+++ client/src/assets/logos/bigLogo.png
