+++ client/src/assets/logos/smallLogo.png
