+++ client/src/assets/styles/globalStyles.scss
@@ -1,0 +1,6 @@
+@import url('https://fonts.googleapis.com/css2?family=Nunito:ital,wght@0,200..1000;1,200..1000&family=Outfit:wght@100..900&display=swap');
+
+@import "variables";
+@import "styles";
+@import "MainPage";
+@import "Navbar";
