+++ client/src/assets/styles/variables.scss
@@ -1,0 +1,6 @@
+$color-primary: #C978FB;
+$color-secondary: #64268A;
+$color-tertiary: #31AFD4;
+$color-quaternary: #093758;
+$color-black: #190624;
+$color-white: #FCFCFC;
