+++ client/src/components/MainPage/MainPage.jsx
@@ -1,22 +1,27 @@
 import React from 'react';
-import bigLogo from './Logos/bigLogo.png'
+import bigLogo from '../../assets/logos/bigLogo.png'
 import Navbar from '../Navbar/Navbar';
+import { useNavigate } from 'react-router-dom';
 
 const MainPage = () => {
+  const navigate = useNavigate();
   return (
-    <section className="main-page">
-      <div className="main-content">
-        <h1>Implementation in seconds</h1>
-        <p>Automatically generate a GraphQL layer over your PostgreSQL database</p>
-        <div className="main-buttons">
-          <button className="upload-btn">Upload Schema</button>
-          <button className="login-btn">Log In</button>
+    <>
+      <Navbar/>
+      <section className="main-page">
+        <div className="main-content">
+          <h1>Implementation in seconds</h1>
+          <p>Automatically generate a GraphQL layer over your PostgreSQL database</p>
+          <div className="main-buttons">
+            <button className="upload-btn" onClick={() => navigate("/signup")}>Sign Up</button>
+            <button className="login-btn">Log In</button>
+          </div>
         </div>
-      </div>
-      <div className="bigLogo">
-        <img src={bigLogo} alt="Main Graphic" />
-      </div>
-    </section>
+        <div className="bigLogo">
+          <img src={bigLogo} alt="Main Graphic" />
+        </div>
+      </section>
+    </>
   );
 };
