+++ client/webpack.config.js
@@ -25,7 +25,6 @@ module.exports = {
       
       {
         test: /\.s?css/,
-      
         use: ['style-loader', 'css-loader', 'sass-loader'],
       },
       {
